add GrpcClientFactory
