+++ tessera-app/src/main/java/com/quorum/tessera/node/grpc/GrpcClientFactory.java
@@ -1,0 +1,26 @@
+package com.quorum.tessera.node.grpc;
+
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+
+public final class GrpcClientFactory {
+
+    private static final ConcurrentHashMap<String, GrpcClient> CLIENTS = new ConcurrentHashMap<>();
+
+    private GrpcClientFactory() {
+
+    }
+
+    private static GrpcClient newClient(final String targetUrl) {
+        final GrpcClient client = new GrpcClient(targetUrl);
+        CLIENTS.put(targetUrl, client);
+        return client;
+    }
+
+    public static GrpcClient getClient(final String targetUrl) {
+        final GrpcClient client = Optional.ofNullable(CLIENTS.get(targetUrl))
+            .orElse(newClient(targetUrl));
+        return client;
+    }
+
+}
